refactor: Condensing cleanupNotifications functions into one file with shared functionality. Also left notes for future updates to use more firebase-idiomatic ways to handle document cleanup
