+++ functions/src/notifications/index.ts
@@ -2,13 +2,15 @@
 import { publishNotifications } from "./publishNotifications"
 import { populateBillHistoryNotificationEvents } from "./populateBillHistoryNotificationEvents"
 import { populateTestimonySubmissionNotificationEvents } from "./populateTestimonySubmissionNotificationEvents"
-import { cleanupNotifications } from "./cleanupNotifications"
+import {
+  cleanupNotifications,
+  httpsCleanupNotifications
+} from "./cleanupNotifications"
 import {
   deliverNotifications,
   httpsDeliverNotifications
 } from "./deliverNotifications"
 import { httpsPublishNotifications } from "./httpsPublishNotifications"
-import { httpsCleanupNotifications } from "./httpsCleanupNotifications"
 import { updateUserNotificationFrequency } from "./updateUserNotificationFrequency"
 
 // Export the functions
